+++ crates/client/db/src/storage_updates.rs
@@ -1,59 +1,70 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use mp_block::state_update::StateUpdateWrapper;
 use mp_contract::class::{
     ClassUpdateWrapper, ContractClassData, ContractClassWrapper, StorageContractClassData, StorageContractData,
 };
-use rayon::prelude::{IntoParallelIterator, IntoParallelRefIterator, ParallelIterator};
+use rayon::prelude::{IntoParallelIterator, ParallelIterator};
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce, PatriciaKey};
 use starknet_api::hash::StarkFelt;
+use starknet_core::types::{DeclaredClassItem, DeployedContractItem, NonceUpdate, ReplacedClassItem, StateUpdate};
 use tokio::task::{self, spawn_blocking};
 
 use crate::storage_handler::{self, DeoxysStorageError, StorageView, StorageViewMut};
 
-pub async fn store_state_update(block_number: u64, state_update: StateUpdateWrapper) -> Result<(), DeoxysStorageError> {
+pub async fn store_state_update(block_number: u64, state_update: StateUpdate) -> Result<(), DeoxysStorageError> {
+    let state_diff = state_update.state_diff.clone();
     let nonce_map: HashMap<ContractAddress, Nonce> = state_update
         .state_diff
         .nonces
         .into_iter()
-        .map(|(contract_address, nonce)| {
+        .map(|NonceUpdate { contract_address, nonce }| {
             (
-                ContractAddress(PatriciaKey(StarkFelt(contract_address.0.to_bytes_be()))),
-                Nonce(StarkFelt(nonce.0.to_bytes_be())),
+                ContractAddress(PatriciaKey(StarkFelt::new_unchecked(contract_address.to_bytes_be()))),
+                Nonce(StarkFelt::new_unchecked(nonce.to_bytes_be())),
             )
         })
         .collect();
 
-    let (result1, result2) = tokio::join!(
+    let (result1, result2, result3) = tokio::join!(
         async move {
             let handler_contract_data = Arc::new(storage_handler::contract_data_mut());
             let handler_contract_data_1 = Arc::clone(&handler_contract_data);
 
-            task::spawn_blocking(move || {
-                state_update
-                    .state_diff
-                    .deployed_contracts
-                    .par_iter()
-                    .chain(state_update.state_diff.replaced_classes.par_iter())
-                    .map(|contract| (ContractAddress(contract.address.into()), ClassHash(contract.class_hash.into())))
-                    .for_each(|(contract_address, class_hash)| {
-                        let previous_nonce =
-                            handler_contract_data_1.get(&contract_address).unwrap().map(|data| data.nonce);
+            let iter_depoyed = state_update.state_diff.deployed_contracts.into_par_iter().map(
+                |DeployedContractItem { address, class_hash }| {
+                    (
+                        ContractAddress(PatriciaKey(StarkFelt::new_unchecked(address.to_bytes_be()))),
+                        ClassHash(StarkFelt::new_unchecked(class_hash.to_bytes_be())),
+                    )
+                },
+            );
+            let iter_replaced = state_update.state_diff.replaced_classes.into_par_iter().map(
+                |ReplacedClassItem { contract_address, class_hash }| {
+                    (
+                        ContractAddress(PatriciaKey(StarkFelt::new_unchecked(contract_address.to_bytes_be()))),
+                        ClassHash(StarkFelt::new_unchecked(class_hash.to_bytes_be())),
+                    )
+                },
+            );
 
-                        handler_contract_data_1
-                            .insert(
-                                contract_address,
-                                StorageContractData {
-                                    class_hash,
-                                    nonce: match nonce_map.get(&contract_address) {
-                                        Some(nonce) => *nonce,
-                                        None => previous_nonce.unwrap_or_default(),
-                                    },
+            task::spawn_blocking(move || {
+                iter_depoyed.chain(iter_replaced).for_each(|(contract_address, class_hash)| {
+                    let previous_nonce = handler_contract_data_1.get(&contract_address).unwrap().map(|data| data.nonce);
+
+                    handler_contract_data_1
+                        .insert(
+                            contract_address,
+                            StorageContractData {
+                                class_hash,
+                                nonce: match nonce_map.get(&contract_address) {
+                                    Some(nonce) => *nonce,
+                                    None => previous_nonce.unwrap_or_default(),
                                 },
-                            )
-                            .unwrap()
-                    });
+                            },
+                        )
+                        .unwrap()
+                });
             })
             .await
             .unwrap();
@@ -69,10 +80,10 @@ pub async fn store_state_update(block_number: u64, state_update: StateUpdateWrap
                     .state_diff
                     .declared_classes
                     .into_par_iter()
-                    .map(|declared_class| {
+                    .map(|DeclaredClassItem { class_hash, compiled_class_hash }| {
                         (
-                            ClassHash(declared_class.class_hash.into()),
-                            CompiledClassHash(declared_class.compiled_class_hash.into()),
+                            ClassHash(StarkFelt::new_unchecked(class_hash.to_bytes_be())),
+                            CompiledClassHash(StarkFelt::new_unchecked(compiled_class_hash.to_bytes_be())),
                         )
                     })
                     .for_each(|(class_hash, compiled_class_hash)| {
@@ -87,11 +98,13 @@ pub async fn store_state_update(block_number: u64, state_update: StateUpdateWrap
                 .commit(block_number)
                 .await
         },
+        async move { storage_handler::block_state_diff().insert(block_number, state_diff) }
     );
 
-    match (result1, result2) {
-        (Err(err), _) => Err(err),
-        (_, Err(err)) => Err(err),
+    match (result1, result2, result3) {
+        (Err(err), _, _) => Err(err),
+        (_, Err(err), _) => Err(err),
+        (_, _, Err(err)) => Err(err),
         _ => Ok(()),
     }
 }